requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-mic-recorder
Pillow
gTTS
groq
